+++ packages/atomic/src/components/facets/facet-segmented-value/facet-segmented-value.tsx
@@ -2,9 +2,8 @@ import {FunctionalComponent, h} from '@stencil/core';
 import {Button} from '../../common/button';
 import {FacetValueProps} from '../facet-common';
 
-export const FacetValueBox: FunctionalComponent<FacetValueProps> = (
-  props,
-  children
+export const FacetSegmentedValue: FunctionalComponent<FacetValueProps> = (
+  props
 ) => {
   const compactCount = new Intl.NumberFormat(props.i18n.language, {
     notation: 'compact',
@@ -20,21 +19,32 @@ export const FacetValueBox: FunctionalComponent<FacetValueProps> = (
   return (
     <li key={props.displayValue}>
       <Button
-        style="outline-bg-neutral"
+        style="square-neutral"
         part="value-box"
         onClick={() => props.onClick()}
-        class={`value-box box-border w-full h-full items-center p-2 group ${
-          props.isSelected ? 'selected' : ''
+        class={`value-box flex box-border h-full items-center p-2 group ${
+          props.isSelected
+            ? 'selected border-primary shadow-inner-primary'
+            : 'hover:border-primary-light focus-visible:border-primary-light'
         }`}
         ariaPressed={props.isSelected.toString()}
         ariaLabel={ariaLabel}
-        ref={props.buttonRef}
       >
-        {children}
+        <span
+          title={props.displayValue}
+          part="value-label"
+          class={`value-label truncate ${
+            props.isSelected
+              ? 'text-primary'
+              : 'group-hover:text-primary-light group-focus:text-primary'
+          }`}
+        >
+          {props.displayValue}
+        </span>
         <span
           title={count}
           part="value-count"
-          class="value-box-count text-neutral-dark truncate w-full text-sm mt-1"
+          class="value-box-count text-neutral-dark truncate pl-1 w-auto mt-0 text-sm"
         >
           {props.i18n.t('between-parentheses', {
             text: compactCount,