+++ packages/atomic/src/components/search-box-suggestions/atomic-search-box-recent-queries/atomic-search-box-recent-queries.tsx
@@ -15,6 +15,7 @@ import {
 } from '../suggestions-common';
 import {once} from '../../../utils/utils';
 import {SafeStorage, StorageItems} from '../../../utils/local-storage-utils';
+import {cleanUpString} from '../../../utils/string-utils';
 
 /**
  * The `atomic-search-box-recent-queries` component can be added as a child of an `atomic-search-box` component, allowing for the configuration of recent query suggestions.
@@ -156,7 +157,7 @@ export class AtomicSearchBoxRecentQueries {
   private renderItem(value: string): SearchBoxSuggestionElement {
     const query = this.bindings.searchBoxController.state.value;
     return {
-      key: `recent-${value}`,
+      key: `recent-${cleanUpString(value)}`,
       query: value,
       content: (
         <div class="flex items-center break-all">